Restructure project folder
