+++ final/src/main/java/com/example/annotation/Column.java
@@ -1,4 +1,4 @@
-package com.example;
+package com.example.annotation;
 import java.lang.annotation.ElementType;
 import java.lang.annotation.Retention;
 import java.lang.annotation.RetentionPolicy;
