+++ final/src/main/java/com/example/connection/ConnectionFactory.java
@@ -1,4 +1,4 @@
-package com.example;
+package com.example.connection;
 
 import java.sql.Connection;
 import java.sql.SQLException;
