+++ final/src/main/java/com/example/connection/DatabaseConfiguration.java
@@ -1,4 +1,4 @@
-package com.example;
+package com.example.connection;
 
 import java.util.Properties;
