+++ final/src/main/java/com/example/connection/DatabaseSession.java
@@ -1,4 +1,4 @@
-package com.example;
+package com.example.connection;
 
 import java.sql.Connection;
 import java.sql.ResultSet;
@@ -7,10 +7,10 @@ import java.sql.Statement;
 
 public class DatabaseSession {
 
-    private Connection connection;
+    private final Connection connection;
 
-    // Constructor that accepts MySQLConnectionFactory instead of connection parameters
-    public DatabaseSession(MySQLConnectionFactory factory) throws SQLException {
+    // Constructor that accepts ConnectionFactory instead of connection parameters
+    public DatabaseSession(ConnectionFactory factory) throws SQLException {
         // Create the connection using the factory
         this.connection = factory.createConnection();
     }
